Refactor: extract deterministic math crate
