+++ crates/warp-geom/src/types/transform.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // © James Ross Ω FLYING•ROBOTS <https://github.com/flyingrobots>
 
-use warp_core::math::{Mat4, Quat, Vec3};
+use warp_math::{Mat4, Quat, Vec3};
 
 /// Rigid transform with non-uniform scale used by broad-phase and shape placement.
 ///
