+++ crates/warp-math/src/bin/gen_sin_qtr_lut.rs
@@ -4,7 +4,7 @@
 //! Developer utility: generate the quarter-wave `sin` LUT for deterministic trig.
 //!
 //! This binary is not used by the runtime. It exists to regenerate the checked-in
-//! LUT in `crates/warp-core/src/math/trig_lut.rs`.
+//! LUT in `crates/warp-math/src/trig_lut.rs`.
 //!
 //! Notes:
 //! - The LUT is generated by sampling `sin` in `f64`, casting to `f32`, and
@@ -21,11 +21,11 @@
 use std::f64::consts::FRAC_PI_2;
 
 fn main() {
-    // MUST match `warp_core::math::trig_lut::SIN_QTR_SEGMENTS` (currently 1024).
+    // MUST match `warp_math::trig_lut::SIN_QTR_SEGMENTS` (currently 1024).
     //
     // Note: `SIN_QTR_SEGMENTS` is `pub(crate)` inside `warp-core`, so this tool
     // cannot reference it directly without widening visibility. Keep this in
-    // lockstep with `crates/warp-core/src/math/trig_lut.rs`.
+    // lockstep with `crates/warp-math/src/trig_lut.rs`.
     const SEGMENTS: usize = 1024;
     const N: usize = SEGMENTS + 1;
     let mut bits: [u32; N] = [0; N];
