+++ crates/warp-math/src/trig.rs
