+++ crates/warp-math/tests/math_additional_tests.rs
@@ -2,7 +2,7 @@
 // © James Ross Ω FLYING•ROBOTS <https://github.com/flyingrobots>
 
 #![allow(missing_docs, clippy::float_cmp)]
-use warp_core::math::{self, Mat4, Quat, Vec3};
+use warp_math::{self as math, Mat4, Quat, Vec3};
 
 fn approx_eq(a: f32, b: f32) {
     let diff = (a - b).abs();
